refactor(wasm): modularize WIT runtime crate
