+++ crates/ironclaw_wasm/src/bindings.rs
@@ -1,0 +1,7 @@
+#![allow(clippy::all)]
+
+wasmtime::component::bindgen!({
+    path: "../../wit/tool.wit",
+    world: "sandboxed-tool",
+    with: {},
+});
